cloud/requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==6.0.3
django-appconf==1.2.0
django-cryptography==1.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
pycparser==3.0
pycryptodome==3.23.0
python-decouple==3.8
sqlparse==0.5.5
